refactor cli input modules
