+++ crates/rulemorph_cli/tests/cli.rs
@@ -18,6 +18,8 @@ include!("cli/preflight.rs");
 
 include!("cli/transform_command.rs");
 
+include!("cli/context.rs");
+
 include!("cli/validate.rs");
 
 include!("cli/limits.rs");
